requirements/local.txt
# Local development dependencies go here
-r base.txt

django-debug-toolbar==1.4

pytest==2.8.2
pytest-django==2.9.1
coverage==4.0.1
